major changes, dtypes back to old fashion, array lists now working fine

+++ include/array_list.h
@@ -5,16 +5,16 @@
 #include <stdlib.h>
 
 typedef struct array_list_s {
-    unsigned long long *array;
-    unsigned long long *times;
-    unsigned long long length;
-    unsigned long long capacity;
+    int *array;
+    int *times;
+    int length;
+    int capacity;
 } array_list;
 
 void array_list_init(array_list **, int);
 void array_list_clean(array_list **, int);
 void array_list_destroy(array_list **, int);
-void array_list_push(array_list **, unsigned long long, unsigned long long, int);
+void array_list_push(array_list **, int, int, int);
 
 
 #endif  /* ARRAY_LIST_H */
